[POM-30] rename package ogloszenia -> ads
