+++ src/main/java/pl/gov/coi/pomocua/ads/OgloszeniaApplication.java
@@ -1,0 +1,12 @@
+package pl.gov.coi.pomocua.ads;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AdsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AdsApplication.class, args);
+	}
+}
