+++ src/test/java/pl/gov/coi/pomocua/ads/OgloszeniaApplicationTests.java
@@ -1,0 +1,13 @@
+package pl.gov.coi.pomocua.ads;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AdsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
