+++ src/test/java/pl/gov/coi/pomocua/ads/work/WorkResourceTest.java
@@ -1,4 +1,4 @@
-package pl.gov.coi.pomocua.ogloszenia.work;
+package pl.gov.coi.pomocua.ads.work;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +8,7 @@ import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import pl.gov.coi.pomocua.ogloszenia.PageableResponse;
+import pl.gov.coi.pomocua.ads.PageableResponse;
 
 import java.util.List;
 import java.util.UUID;
